+++ src/navigations/bottomTabs/MainTapNavigator.tsx
@@ -18,19 +18,14 @@ import MainTaskScreen from '../../screens/MainTaskScreen/MainTaskScreen';
 import ProfileScreen from '../../screens/ProfileScreen/ProfileScreen';
 import RegisterNotification from '../../components/Modal/RegisterNotification';
 import {TabActions} from '@react-navigation/native';
-import RegisterFailedNotification from '../../components/Modal/RegisterFailedNotification';
 import RegisterFailedModal from '../../components/Modal/RegisterFailedModalNotification';
 import Toast from 'react-native-toast-message';
 import {responsiveHeigth, responsiveWidth} from '../../function/responsive';
 import * as RootNavigation from '../../navigations/RootNavigation';
-import {SheetManager} from 'react-native-actions-sheet';
 import {ActionContext} from '../../../App';
 import {dialCall} from '../../function/utility';
 import RewardScreen from '../../screens/RewardScreen';
 import MissionScreen from '../../screens/MissionScreen';
-import Draggable from 'react-native-draggable';
-import AsyncStorage from '@react-native-async-storage/async-storage';
-import {Campaign} from '../../datasource/CampaignDatasource';
 
 export type TabNavigatorParamList = {
   mission: undefined;
@@ -38,6 +33,7 @@ export type TabNavigatorParamList = {
   myTask: undefined;
   reward: undefined;
   profile: undefined;
+  CampaignScreen: undefined;
 };
 const Tab = createBottomTabNavigator();